+++ src/users/test/user.e2e.spec.ts
@@ -7,6 +7,8 @@ import {initializeTransactionalContext} from "typeorm-transactional";
 import {DataSource} from "typeorm";
 import {UserTokenFactory} from "../../common/testSetup/user/userTokenFactory";
 import {UserFinder} from "../../common/testSetup/user/userFinder";
+import * as path from 'path';
+
 
 describe('UserController (E2E)', () => {
     let app: INestApplication;
@@ -14,7 +16,12 @@ describe('UserController (E2E)', () => {
 
     beforeAll(async () => {
         initializeTransactionalContext();
-        dotenv.config();
+        dotenv.config({
+            path: path.resolve(
+                process.env.NODE_ENV === 'product' ? '.env.product' :
+                    process.env.NODE_ENV === 'develop' ? '.env.develop' : '.env.local'
+            )
+        });
         const moduleRef: TestingModule = await Test.createTestingModule({
             imports: [AppModule]
         }).compile();
@@ -27,7 +34,9 @@ describe('UserController (E2E)', () => {
     })
 
     beforeEach(async () => {
-        await dataSource.dropDatabase();
+        if (process.env.NODE_ENV === 'develop' || process.env.NODE_ENV === 'local') {
+            await dataSource.dropDatabase();
+        }
         await dataSource.synchronize();
     })
